requirements.txt
requests
python-dotenv
pandas
geopy
plotly
dash
pytest
